refactor(extractor): inline abxxx/metadata.rs into mod.rs
After the KVS promotion in 919df6b, abxxx/metadata.rs held a single
30-line slug_search_endpoint helper, no metadata at all — a misnomer
left over from when the file owned the rich lookup parsing that's now
in base/kvs/api.rs.

Inlined into abxxx/mod.rs next to videofile_endpoint(), where the
matching test also moves. No behaviour change; the helper is private
and called from one place.
